refactor : 상수클래스 패키지 변경
