+++ src/main/java/bridge/domain/BridgeGame.java
@@ -4,8 +4,8 @@ import bridge.util.validation.GameValidator;
 
 import java.util.List;
 
-import static bridge.util.GameRule.COMMAND_QUIT;
-import static bridge.util.GameRule.COMMAND_RETRY;
+import static bridge.util.constants.GameRule.COMMAND_QUIT;
+import static bridge.util.constants.GameRule.COMMAND_RETRY;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
@@ -23,7 +23,7 @@ public class BridgeGame {
 
     public boolean judge(String playerMoving) {
         GameValidator.validateMovingInput(playerMoving);
-        return player.compare(playerMoving, bridge);
+        return player.isMovableBy(playerMoving, bridge);
     }
 
     public void move(String playerMoving) {
