+++ src/main/java/bridge/util/constants/Exception.java
@@ -1,4 +1,4 @@
-package bridge.util.validation;
+package bridge.util.constants;
 
 public class Exception {
     public static final String ERROR_MESSAGE_FOR_NOT_NUMBER = "[ERROR] 숫자를 입력해주세요.";
