+++ src/main/java/bridge/util/validation/GameValidator.java
@@ -1,12 +1,13 @@
 package bridge.util.validation;
 
 import bridge.domain.Bridge;
+import bridge.util.constants.Exception;
 import java.util.List;
 
-import static bridge.util.GameRule.BRIDGE_LENGTH_MAX;
-import static bridge.util.GameRule.BRIDGE_LENGTH_MIN;
-import static bridge.util.GameRule.COMMAND_QUIT;
-import static bridge.util.GameRule.COMMAND_RETRY;
+import static bridge.util.constants.GameRule.BRIDGE_LENGTH_MAX;
+import static bridge.util.constants.GameRule.BRIDGE_LENGTH_MIN;
+import static bridge.util.constants.GameRule.COMMAND_QUIT;
+import static bridge.util.constants.GameRule.COMMAND_RETRY;
 
 public class GameValidator {
     public static void validateBridgeSize(int size) {
@@ -32,7 +33,7 @@ public class GameValidator {
     }
 
     public static void validatePlayerLocation(int location, List<String> bridge) {
-        if(location == bridge.size()) {
+        if(location >= bridge.size()) {
             throw new IllegalArgumentException(Exception.ERROR_MESSAGE_OUT_RANGE_OF_BRIDGE);
         }
     }
