+++ src/main/java/bridge/util/validation/ViewValidator.java
@@ -1,5 +1,7 @@
 package bridge.util.validation;
 
+import bridge.util.constants.Exception;
+
 public class ViewValidator {
     public static void isNumber(String input) {
         if (input.chars().allMatch(Character::isDigit)) {
